Use identity-did for more functionality
