+++ identity-core/src/crypto/mod.rs
@@ -5,5 +5,17 @@
 
 mod key;
 pub mod merkle_tree;
+mod proof;
+mod signature;
 
-pub use self::{key_impl::*, key_pair::*};
+pub use self::key::KeyPair;
+pub use self::key::PublicKey;
+pub use self::key::SecretKey;
+pub use self::proof::JcsEd25519Signature2020;
+pub use self::signature::SigName;
+pub use self::signature::SigSign;
+pub use self::signature::SigVerify;
+pub use self::signature::Signature;
+pub use self::signature::SignatureData;
+pub use self::signature::SignatureOptions;
+pub use self::signature::SignatureValue;
