+++ identity-core/src/crypto/proof/mod.rs
@@ -6,4 +6,4 @@
 
 mod jcsed25519signature2020;
 
-pub use self::jcsed25519signature2020::*;
+pub use self::jcsed25519signature2020::JcsEd25519Signature2020;
